Move test_two_columns_with_different_dtypes to nonreg tests
